remove some unused code
