--- Sources/APNSwift/TEMP.swift
@@ -1,1 +1,0 @@
-extension String: Error { }
